simplify account utils to one file, remove duplicate function defitions, fix imports, fix comment, fix init values
