+++ qa/SidechainTestFramework/account/ac_make_raw_tx.py
@@ -1,30 +1,8 @@
-import os
 import subprocess
 import logging
+from SidechainTestFramework.account.ac_utils import install_npm_packages, get_cwd
 
 nodeModulesInstalled = False
-cwd = None
-
-
-def get_cwd():
-    global cwd
-    if cwd is None:
-        cwd = os.environ.get("SIDECHAIN_SDK") + "/qa/SidechainTestFramework/account/smart_contract_resources"
-    return cwd
-
-
-def install_npm_packages():
-    global nodeModulesInstalled
-    if not nodeModulesInstalled:
-        logging.info("Installing node packages...")
-        logging.info("The first time this runs on your machine, it can take a few minutes...")
-        proc = subprocess.run(
-            ["yarn", "install", "--quiet", "--non-interactive", "--frozen-lockfile"],
-            cwd=get_cwd(),
-            stdout=subprocess.PIPE)
-        proc.check_returncode()
-        logging.info("Done!")
-        nodeModulesInstalled = True
 
 
 def legacy(*, to: str,
@@ -99,9 +77,6 @@ def eip1559(*, to: str,
         output = output.split(' ')[1].strip()
         logging.info("Returned raw transaction:", '"' + output + '"')
         return output
-
-
-# def load_contract_abi(contractPath: str)
 
 
 if __name__ == '__main__':
