+++ qa/SidechainTestFramework/account/ac_smart_contract_compile.py
@@ -2,30 +2,9 @@ import os
 import subprocess
 import logging
 
-nodeModulesInstalled = False
+from SidechainTestFramework.account.ac_utils import install_npm_packages, get_cwd
+
 sigHashesGenerated = False
-cwd = None
-
-
-def get_cwd():
-    global cwd
-    if cwd is None:
-        cwd = os.environ.get("SIDECHAIN_SDK") + "/qa/SidechainTestFramework/account/smart_contract_resources"
-    return cwd
-
-
-def install_npm_packages():
-    global nodeModulesInstalled
-    if not nodeModulesInstalled:
-        logging.info("Installing node packages...")
-        logging.info("The first time this runs on your machine, it can take a few minutes...")
-        proc = subprocess.run(
-            ["yarn", "install", "--quiet", "--non-interactive", "--frozen-lockfile"],
-            cwd=get_cwd(),
-            stdout=subprocess.PIPE)
-        proc.check_returncode()
-        logging.info("Done!")
-        nodeModulesInstalled = True
 
 
 def compile_smart_contracts():
@@ -52,9 +31,6 @@ def compile_smart_contracts():
 def prepare_resources():
     install_npm_packages()
     compile_smart_contracts()
-
-
-# def load_contract_abi(contractPath: str)
 
 
 if __name__ == '__main__':
