+++ qa/SidechainTestFramework/account/ac_use_smart_contract.py
@@ -8,11 +8,8 @@ from SidechainTestFramework.account.ac_smart_contract_compile import prepare_res
 import os
 from dataclasses import dataclass
 
-from SidechainTestFramework.account.address_util import format_eoa, format_evm
-from SidechainTestFramework.account.evm_util import CallMethod
+from SidechainTestFramework.account.ac_utils import CallMethod, format_eoa, format_evm, ensure_nonce, ensure_chain_id
 from SidechainTestFramework.account.mk_contract_address import mk_contract_address
-from enum import Enum
-import SidechainTestFramework.account.ac_make_raw_tx as make_raw_tx
 
 
 class EvmExecutionError(RuntimeError):
@@ -46,7 +43,6 @@ class SmartContract:
     """This class represents a type of smart contract. Function calls and deployments etc. are taken care of in this
     implementation, specifically the encoding and sending of transactions via http or rpc"""
 
-    # TODO auto find nonce once implemnted if None
     # TODO auto parse output and stuff once receipts and static calls work
     def __init__(self, contract_path: str):
         """The constructor argument is a unique name of a smart contract.
@@ -104,8 +100,8 @@ class SmartContract:
 
         """
         data = self.raw_encode_call(functionName, *args)
-        nonce = self.__ensure_nonce(node, fromAddress, nonce, tag)
-        chain_id = self.__ensure_chain_id(node)
+        nonce = ensure_nonce(node, fromAddress, nonce, tag)
+        chain_id = ensure_chain_id(node)
         response: Any = ''
         if call_method is CallMethod.RPC_LEGACY:
             response = node.rpc_eth_signTransaction(
@@ -129,7 +125,7 @@ class SmartContract:
         raise RuntimeError("Something went wrong, see {}".format(str(response)))
 
     def static_call(self, node, functionName: str, *args, fromAddress: str, nonce: int = None, toAddress: str,
-                    gasLimit: int = 0, gasPrice: int = 0, value: int = 0, tag: str = 'latest'):
+                    gasLimit: int = None, gasPrice: int = None, value: int = 0, tag: str = 'latest'):
         """Calls a function in read-only mode and returns the data if applicable
 
                Parameters:
@@ -137,8 +133,8 @@ class SmartContract:
                    functionName (str):   The name of the function to call. Has to include parentheses and params without spaces (e.g. `someFunction(uint256,string)`
                    *args: the arguments to call the function with, in the correct order.
                    fromAddress: the sender address to use.
-                   toAddress: the address of the smart contract instance to use
                    nonce (optional): The nonce to use. If not given, the nonce will be retrieved via rpc
+                   toAddress: the address of the smart contract instance to use
                    gasLimit: The gasLimit to use
                    gasPrice: The gasPrice to use
                    value: The value to use
@@ -150,13 +146,13 @@ class SmartContract:
         request = {
             "from": format_evm(fromAddress),
             "to": format_evm(toAddress),
-            "nonce": self.__ensure_nonce(node, fromAddress, nonce, tag),
+            "nonce": ensure_nonce(node, fromAddress, nonce, tag),
             "value": value,
             "data": self.raw_encode_call(functionName, *args)
         }
-        if gasLimit > 0:
+        if gasLimit is not None:
             request["gas"] = gasLimit
-        if gasPrice > 0:
+        if gasPrice is not None:
             request["gasPrice"] = gasPrice
         response = node.rpc_eth_call(request, tag)
         if 'result' in response:
@@ -180,7 +176,7 @@ class SmartContract:
                      tag: str = 'latest'):
         request = {
             "from": format_evm(fromAddress),
-            "nonce": self.__ensure_nonce(node, fromAddress, nonce, tag),
+            "nonce": ensure_nonce(node, fromAddress, nonce, tag),
             "value": value
         }
         if functionName == 'constructor':
@@ -208,24 +204,25 @@ class SmartContract:
     def deploy(self, node, *args, call_method: CallMethod = CallMethod.RPC_LEGACY, fromAddress: str, nonce: int = None,
                gasLimit: int, gasPrice: int = 765630000, maxFeePerGas: int = 765630000, maxPriorityFeePerGas: int = 10,
                value: int = 0, tag: str = 'latest'):
-        """Calls a function in read-only mode and returns the data if applicable
+        """Deploys contract via eth_sendRawTransaction, returns the transaction hash and contract address if applicable
             :param node: The side chain node to use for rpc/http
+            :param args: The smart contract constructor arguments
             :param call_method: The tx type
-            :param value:  The value to use
-            :param tag: The block tag to use when calling rpc methods
-            :param nonce:  The nonce to use. If not given, the nonce will be retrieved via rpc
             :param fromAddress: the sender address to use
-            :param maxPriorityFeePerGas: the maxPriorityFeePerGas in eip1559
+            :param nonce:  The nonce to use. If not given, the nonce will be retrieved via rpc
             :param gasLimit: The gasLimit to use
             :param gasPrice:  The gasPrice to use in legacy/eip155 transactions
             :param maxFeePerGas:  the maxFeePerGas in eip1559
+            :param maxPriorityFeePerGas: the maxPriorityFeePerGas in eip1559
+            :param value:  The value to use
+            :param tag: The block tag to use when calling rpc methods
 
         Returns:
             A tuple with the tx_hash and the precomputed address of the smart contract
 """
-        nonce = self.__ensure_nonce(node, format_evm(fromAddress), nonce, tag)
+        nonce = ensure_nonce(node, format_evm(fromAddress), nonce, tag)
         data = self.Bytecode
-        chain_id = self.__ensure_chain_id(node)
+        chain_id = ensure_chain_id(node)
 
         if 'constructor' in self.Functions:
             if len(args) != len(self.Functions['constructor'].inputs):
@@ -283,10 +280,10 @@ class SmartContract:
             return self.Functions[function_name].encode(*args)
 
     def raw_decode_call_result(self, function_name: str, data):
-        """Can be used to decode *args for function_name
+        """Can be used to decode data for function_name
             Params:
                 function_name: The name of the function including parentheses and input types
-                *args: the arguments to decode
+                data: the arguments to decode
 
             Returns:
                 A tuple of the decoded values
@@ -341,7 +338,7 @@ class SmartContract:
 
     @staticmethod
     def __make_eip1559_sign_payload(*, from_addr: str, to: str = None, nonce: int, gas: int, value: int,
-                                    data: str = None, max_priority_fee_per_gas: int = None, max_fee_per_gas: int = None,
+                                    data: str = '0x', max_priority_fee_per_gas: int = None, max_fee_per_gas: int = None,
                                     chain_id: str):
         r = {
             "type": 2,
@@ -410,30 +407,8 @@ class SmartContract:
                 "Contract name is not unique, please change the names of the contracts so they are unique")
         return sol_files[0]
 
-    @staticmethod
-    def __ensure_nonce(node, address, nonce, tag='latest'):
-        on_chain_nonce = int(node.rpc_eth_getTransactionCount(format_evm(address), tag)['result'], 16)
-        if nonce is None:
-            nonce = on_chain_nonce
-        return nonce
-
-    @staticmethod
-    def __ensure_chain_id(node):
-        return node.rpc_eth_chainId()['result']
-
 
 if __name__ == '__main__':
-    # logging.info(get_cwd())
-    # try:
-    #     SmartContract("path")
-    # except RuntimeError as err:
-    #     pass
-    # try:
-    #     SmartContract(
-    #         "contracts/ExampleERC20Contract.sol")
-    # except RuntimeError as err:
-    #     pass
-    # SmartContract("contracts/ExampleERC20")
     logging.info("Loading example contract and testing encoding")
     sc = SmartContract("StorageTestContract.sol")
     logging.info(sc)
