+++ qa/sc_evm_test_block_bloom_filter.py
@@ -4,8 +4,8 @@ from eth_utils import keccak
 from eth_bloom import BloomFilter
 from decimal import Decimal
 
-from SidechainTestFramework.account.ac_use_smart_contract import SmartContract, EvmExecutionError
-from SidechainTestFramework.account.address_util import format_evm
+from SidechainTestFramework.account.ac_use_smart_contract import SmartContract
+from SidechainTestFramework.account.ac_utils import format_evm, CallMethod, eoa_transaction
 from SidechainTestFramework.sc_boostrap_info import SCNodeConfiguration, SCCreationInfo, MCConnectionInfo, \
     SCNetworkConfiguration, LARGE_WITHDRAWAL_EPOCH_LENGTH
 from SidechainTestFramework.sc_test_framework import SidechainTestFramework
@@ -14,8 +14,6 @@ from test_framework.util import assert_equal, assert_true, start_nodes, websocke
 from SidechainTestFramework.scutil import bootstrap_sidechain_nodes, start_sc_nodes, \
     is_mainchain_block_included_in_sc_block, check_mainchain_block_reference_info, AccountModelBlockVersion, \
     EVM_APP_BINARY, generate_next_blocks, generate_next_block, DEFAULT_EVM_APP_GENESIS_TIMESTAMP_REWIND
-from SidechainTestFramework.account.evm_util import CallMethod
-from SidechainTestFramework.account.eoa_util import eoa_transaction
 
 global_call_method = CallMethod.RPC_EIP155
