+++ qa/sc_evm_test_contract_contract_deployment_and_interaction.py
@@ -6,7 +6,7 @@ from decimal import Decimal
 from eth_utils import to_checksum_address
 
 from SidechainTestFramework.account.ac_use_smart_contract import SmartContract
-from SidechainTestFramework.account.address_util import format_evm
+from SidechainTestFramework.account.ac_utils import format_evm
 from SidechainTestFramework.sc_boostrap_info import SCNodeConfiguration, SCCreationInfo, MCConnectionInfo, \
     SCNetworkConfiguration, LARGE_WITHDRAWAL_EPOCH_LENGTH
 from SidechainTestFramework.sc_test_framework import SidechainTestFramework
